Move SQL stuff into sql.rs
