+++ doc/get.md
@@ -2,14 +2,14 @@
 
 You can read rows from tables using `nanobot get`:
 
-```console tesh-session="test"
+```console tesh-session="get"
 $ nanobot init
 Initialized a Nanobot project
 $ nanobot get table
-table     path                     description                         type
-table     src/schema/table.tsv     All of the tables in this project.  table
-column    src/schema/column.tsv    Columns for all of the tables.      column
-datatype  src/schema/datatype.tsv  Datatypes for all of the columns    datatype
+table     path                     type      description
+table     src/schema/table.tsv     table     All of the tables in this project.
+column    src/schema/column.tsv    column    Columns for all of the tables.
+datatype  src/schema/datatype.tsv  datatype  Datatypes for all of the columns
 ```
 
 This reads the first 100 rows of the 'table' table
@@ -20,26 +20,26 @@ For machine-readability use `--format json`.
 The output is designed to match [PostgREST](https://postgrest.org).
 Piping the output through `jq` makes it easier to read:
 
-```console
+```console tesh-session="get"
 $ nanobot get table --format json | jq
 [
   {
     "table": "table",
     "path": "src/schema/table.tsv",
-    "description": "All of the tables in this project.",
-    "type": "table"
+    "type": "table",
+    "description": "All of the tables in this project."
   },
   {
     "table": "column",
     "path": "src/schema/column.tsv",
-    "description": "Columns for all of the tables.",
-    "type": "column"
+    "type": "column",
+    "description": "Columns for all of the tables."
   },
   {
     "table": "datatype",
     "path": "src/schema/datatype.tsv",
-    "description": "Datatypes for all of the columns",
-    "type": "datatype"
+    "type": "datatype",
+    "description": "Datatypes for all of the columns"
   }
 ]
 ```
