+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod get;
 pub mod serve;
+pub mod sql;
